Move the LycanService into the runner module
