+++ editor/modules/runner.module.ts
@@ -1,9 +1,9 @@
 import {NgModule} from '@angular/core';
 import {SharedModule} from '../components/shared.module';
-import {LycanService} from '../services/lycan';
 
 import {GameCanvas} from './runner/game-canvas.component';
 import {GameState} from './runner/game-state';
+import {LycanService} from './runner/lycan.service';
 
 
 @NgModule({
